Change blocks to be audited by their prior validator set
Instead of their posterior validator set.

The prior validator set needs to be used by GRANDPA for validator set
handover to work properly. Specifically, the first block in an epoch
must be finalized by the previous validator set. This ensures that there
is an unequivocal starting point for the next validator set, and there
is no possibility of accidentally "unfinalizing" a block.

The validator set used for auditing should match the set used by
GRANDPA, as validators use auditing results to determine how they vote
in GRANDPA. Hence the change to use the prior validator set for
auditing.